:art: cleaned up http client creation
